Fresh start: a full C rewrite
I have decided to rewrite ``sprd'' entirely in C. The current
implimentation is in C++, with an extremely fragmented file
structure. The new rewrite will be faster, more memory-efficient, and
easier to read and contribute to. This is the last commit that will
function of the C++ codebase, if you want to use it for some reason.

To the successful rewrite!