sleep: add next_chunk()
Impliment next_chunk(), which takes a C string with a length parameter
and returns address points for the start and end of the next
chunk. next_chunk() does not use null-termination because it is
designed to work with files directly mmap-ed from disk, rather than
read into a ``normal'' string format.